[refactor] game package 생성, 코드 리팩토링
domain.game 게임 진행에 관련된 클래스를 담는 패키지 생성
ALL 패키지 변경에 따른 코드 리팩토링
