+++ src/main/java/domain/card/CardDeck.java
@@ -5,10 +5,11 @@ import java.util.List;
 import java.util.Stack;
 
 public class CardDeck {
-    private final Stack<Card> cards = new Stack<>();
+    private final Stack<Card> cards;
 
     CardDeck(List<Card> cards) {
         Collections.shuffle(cards);
+        this.cards = new Stack<>();
         this.cards.addAll(cards);
     }
