+++ src/main/java/domain/game/Moneys.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import java.util.HashMap;
 import java.util.Map;
