+++ src/main/java/domain/game/PlayerAnswer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import java.util.Arrays;
