+++ src/test/java/domain/card/CardDeckGeneratorForTest.java
@@ -4,11 +4,11 @@ import java.util.Arrays;
 import java.util.List;
 
 public class CardDeckGeneratorForTest {
-    public static CardDeck create(List<Card> cards){
+    public static CardDeck create(List<Card> cards) {
         return new CardDeck(cards);
     }
 
-    public static CardDeck createByOneCard(Card card){
+    public static CardDeck createByOneCard(Card card) {
         return new CardDeck(Arrays.asList(card));
     }
 }
